chore: moved data enum and added enum supported types def
