+++ beet/src/beets/composites.ts
@@ -1,5 +1,4 @@
 import { strict as assert } from 'assert'
-import { u8 } from './numbers'
 import {
   assertFixedSizeBeet,
   Beet,
@@ -149,51 +148,13 @@ export function coption<T, V = T>(inner: Beet<T, V>): FixableBeet<COption<T>> {
 }
 
 /**
- * Represents an {@link Enum} type which contains data.
- *
- * @template Kind the enum variant, i.e. `Color.Red`
- * @template Data the data value, i.e. '#f00'
- *
- * @category beet/composite
- */
-export type DataEnum<Kind, Data> = { kind: Kind & number; data: Data }
-/**
- * De/Serializes an {@link Enum} that contains a type of data, i.e. a {@link Struct}.
- * The main difference to a Rust enum is that the type of data has to be the
- * same for all enum variants.
- *
- * @template T inner enum data type
- *
- * @param inner the De/Serializer for the data type
- *
- * @category beet/composite
- */
-export function dataEnum<Kind, Data>(
-  inner: FixedSizeBeet<Data>
-): FixedSizeBeet<DataEnum<Kind, Data>> {
-  return {
-    write: function (buf: Buffer, offset: number, value: DataEnum<Kind, Data>) {
-      u8.write(buf, offset, value.kind)
-      inner.write(buf, offset + 1, value.data)
-    },
-
-    read: function (buf: Buffer, offset: number): DataEnum<Kind, Data> {
-      const kind = u8.read(buf, offset) as DataEnum<Kind, Data>['kind']
-      const data = inner.read(buf, offset + 1)
-      return { kind, data }
-    },
-    byteSize: 1 + inner.byteSize,
-    description: `DataEnum<${inner.description}>`,
-  }
-}
-/**
  * @category TypeDefinition
  */
 export type CompositesExports = keyof typeof import('./composites')
 /**
  * @category TypeDefinition
  */
-export type CompositesTypeMapKey = 'option' | 'enum'
+export type CompositesTypeMapKey = 'option'
 /**
  * @category TypeDefinition
  */
@@ -215,14 +176,6 @@ export const compositesTypeMap: CompositesTypeMap = {
     isFixable: true,
     sourcePack: BEET_PACKAGE,
     ts: 'COption<Inner>',
-    arg: BEET_TYPE_ARG_INNER,
-    pack: BEET_PACKAGE,
-  },
-  enum: {
-    beet: 'dataEnum',
-    isFixable: false,
-    sourcePack: BEET_PACKAGE,
-    ts: 'DataEnum<Kind, Inner>',
     arg: BEET_TYPE_ARG_INNER,
     pack: BEET_PACKAGE,
   },
