+++ beet/src/beets/enums.ts
@@ -1,15 +1,20 @@
-import { FixedSizeBeet } from '../types'
+import {
+  BEET_PACKAGE,
+  BEET_TYPE_ARG_INNER,
+  FixedSizeBeet,
+  SupportedTypeDefinition,
+} from '../types'
 import { u8 } from './numbers'
 import { strict as assert } from 'assert'
 
 type Enum = { [key: number]: string | number } | string | number
 
 /**
- * De/serializer for enums with up to 255 data less variants.
+ * De/serializer for enums with up to 255 less variants which have no data.
  *
  * @param enumType type of enum to process, i.e. Color or Direction
  */
-export function simpleEnum<T extends Enum, V = T>(
+export function fixedScalarEnum<T extends Enum, V = T>(
   enumType: T
 ): FixedSizeBeet<T, V> {
   return {
@@ -42,4 +47,85 @@ export function simpleEnum<T extends Enum, V = T>(
     byteSize: u8.byteSize,
     description: 'Enum',
   }
+}
+
+/**
+ * Represents an {@link Enum} type which contains fixed size data.
+ *
+ * @template Kind the enum variant, i.e. `Color.Red`
+ * @template Data the data value, i.e. '#f00'
+ *
+ * @category beet/composite
+ */
+export type DataEnum<Kind, Data> = { kind: Kind & number; data: Data }
+/**
+ * De/Serializes an {@link Enum} that contains a type of data, i.e. a {@link Struct}.
+ * The main difference to a Rust enum is that the type of data has to be the
+ * same for all enum variants.
+ *
+ * @template T inner enum data type
+ *
+ * @param inner the De/Serializer for the data type
+ *
+ * @category beet/composite
+ */
+export function dataEnum<Kind, Data>(
+  inner: FixedSizeBeet<Data>
+): FixedSizeBeet<DataEnum<Kind, Data>> {
+  return {
+    write: function (buf: Buffer, offset: number, value: DataEnum<Kind, Data>) {
+      u8.write(buf, offset, value.kind)
+      inner.write(buf, offset + 1, value.data)
+    },
+
+    read: function (buf: Buffer, offset: number): DataEnum<Kind, Data> {
+      const kind = u8.read(buf, offset) as DataEnum<Kind, Data>['kind']
+      const data = inner.read(buf, offset + 1)
+      return { kind, data }
+    },
+    byteSize: 1 + inner.byteSize,
+    description: `DataEnum<${inner.description}>`,
+  }
+}
+
+/**
+ * @category TypeDefinition
+ */
+export type EnumsExports = keyof typeof import('./enums')
+/**
+ * @category TypeDefinition
+ */
+export type EnumsTypeMapKey = 'fixedScalarEnum' | 'dataEnum'
+/**
+ * @category TypeDefinition
+ */
+export type EnumsTypeMap = Record<
+  EnumsTypeMapKey,
+  SupportedTypeDefinition & { beet: EnumsExports }
+>
+
+/**
+ * Maps composite beet exports to metadata which describes in which package it
+ * is defined as well as which TypeScript type is used to represent the
+ * deserialized value in JavaScript.
+ *
+ * @category TypeDefinition
+ */
+export const enumsTypeMap: EnumsTypeMap = {
+  fixedScalarEnum: {
+    beet: 'fixedScalarEnum',
+    isFixable: false,
+    sourcePack: BEET_PACKAGE,
+    ts: '<TypeName>',
+    arg: BEET_TYPE_ARG_INNER,
+    pack: BEET_PACKAGE,
+  },
+  dataEnum: {
+    beet: 'dataEnum',
+    isFixable: false,
+    sourcePack: BEET_PACKAGE,
+    ts: 'DataEnum<Kind, Inner>',
+    arg: BEET_TYPE_ARG_INNER,
+    pack: BEET_PACKAGE,
+  },
 }
